Fix reading of prefix list
The function getS3ObjectList() can accept a list of S3 object
prefixes to filter the list of objects it creates.  This commit
fixes several issues with this code.  First, the comparison of
prefix to object key was being done incorrectly.  Change the
code to use the correct part of the key.

Second, the prefixes had newlines at the end, preventing the
comparison from matching.  Use rstrip() to remove newlines and
any other trailing white space from the prefixes.

Finally, catch the error if open() can't find the prefix file
and return the correct value.