Minor formatting changes
Change the messages at the start and end of the program run
to use better formatting.  Also, print the number of jobs
actually queued at the end.  If there aren't enough files with
the specified prefix in the S3 bucket, submit_multi might not
submit as many jobs as the user asked for.