Add options to submit_multi
Add a number of options to submit_mulit.py.  Chiefly,
allow submit_multi to read a file of S3 URLs and submit
jobs based on them.  Get rid of the prefix file in favor
of the --prefix option.  Refactor code into a main()
function.  Change the order number to use a random
alphabetic code instead of starting at a random number.

In sqsPublisherHelpers.py, add a getListFromFile()
function to handle submit_multi's job file option.
Change getS3ObjectList() to take a prefix list and
use the filter option to find files to submit.

In sqsPublisher.py, adjust code for changes to getS3ObjectList().