Add arguments to submit_multi.py
Add --batch-command and --blacklist arguments to submit_multi.py.

If present, --batch-command is passed through to espa_submit.py.

If --blacklist is set, it should name a file of product IDs that
are to be excluded from processing.  This is a way to avoid
submitting jobs for scenes that we know will fail.