Don't submit jobs we know will fail
Change submit_multi and supporting code to avoid two types of
scenes we know we can't process.  The two types are scenes on
ascending (nighttime) paths, where the solar zenith angle is
out of range, and scenes with L1GT in the product ID, which
don't have a processor implemented.

In submit_multi.py:
- Add a --list-only option to produce a list of scenes
  without submitting them to a queue,
- Set the correct default value for --prefix,

In sqsPublisherHelpers.py, split the input product ID,
check the path/row, look for L1GT in the file name, and
skip scenes that meet either criterion.