Check AWS region when creating boto3 clients
The AWS boto3 library requires the region to be
set when clients are created.  The two methods for
users to specify the AWS region are an AWS configuration
file (.aws/config) and the environment variable
AWSRegion.  If a user doesn't have an AWS configuration
file, but uses AWSRegion, submit_array.py and
submit_multi.py will fail when they try to create boto3
clients.

Always check for the AWSRegion environment variable and,
if it exists, pass it to boto3 in the client() call.