Don't ignore --prefix flag
Ips_submit_multi.py has a --prefix flag, but it fails to use
the value that is specified.  Create a prefix list as necessary
and pass it to the function that gets the object list from the
S3 bucket.

Change the batch job submit call to specifiy the full pathname
to ips-batch-worker.

Add some comments.  Add a message about gathering the S3 object list.