Add BPF file name gathering to ips_submit_multi.py
The main change in this commit is to add code to gather
the names of all the BPF files in the auxiliary data bucket
and write them to files, one for OLI BPF files, the other for
TIRS BPF files.  We upload the files to the job bucket for
later user by find_aux_files.py.  The point of this change is
that find_aux_files.py currently gathers the list of about
90,000 BPF files when it needs to find a BPF.  When a large
number of jobs start at the same time, this puts a strain on
the S3 bucket and can cause data transfer failures.

Remove the --package option, which never should have been there
in the first place.  Clean up the argument parsing and validation
code.  Get the satellite ID and use it to make sure we do the
right thing for different satellites (mostly L8 vs. pre-L8).  Use
os.path functions instead of doing direct pathname manipulation.